requirements.txt
pandas==1.5.3
phidata==2.5.32
psycopg2
psycopg[binary]